Cleanup the Tracing class and move setup as instance method of TracingConfig
